Greatly simplified DiningVenue parsing logic by depending completely on custom Codable implementations
